Extract type definitions from Zod and remove from types.ts. Refactoring and usages.
