+++ packages/reassure-compare/src/output/console.ts
@@ -1,4 +1,4 @@
-import type { AddedEntry, CompareResult, CompareEntry, RemovedEntry, PerformanceMetadata } from '../types';
+import type { AddedEntry, CompareResult, CompareEntry, RemovedEntry } from '../types';
 import {
   formatCount,
   formatDuration,
@@ -6,6 +6,7 @@ import {
   formatRenderCountChange,
   formatRenderDurationChange,
 } from '../utils/format';
+import type { PerformanceMetadata } from '../utils/validate';
 
 export function printToConsole(data: CompareResult) {
   // No need to log errors or warnings as these were be logged on the fly
